lol, wtf, it didnt work

+++ libft-1.3/includes/num.h
@@ -6,7 +6,7 @@
 /*   By: gcros <gcros@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/17 19:51:40 by gcros             #+#    #+#             */
-/*   Updated: 2024/03/07 14:25:44 by gcros            ###   ########.fr       */
+/*   Updated: 2024/09/17 12:51:33 by gcros            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@
 
 int		ft_atoi(const char *nptr);
 long	ft_atol(const char *nptr);
-float	ft_maxf(float n1, float n2);
-float	ft_minf(float n1, float n2);
-float	ft_absf(float n);
+double	ft_maxf(double n1, double n2);
+double	ft_minf(double n1, double n2);
+double	ft_absf(double n);
 int		ft_abs(int n);
 int		ft_min(int n1, int n2);
 int		ft_max(int n1, int n2);
